Fix up import_geographies and rename it to import_shapes
